+++ apps/wing-console/console/app/assets/openChrome.applescript
