+++ apps/wing-console/console/app/package.json
@@ -7,12 +7,12 @@
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "files": [
-    "dist",
-    "scripts"
+    "assets",
+    "dist"
   ],
   "scripts": {
-    "preview": "node scripts/preview.mjs",
-    "dev": "tsx watch scripts/dev.mjs",
+    "preview": "tsx scripts/preview.ts",
+    "dev": "tsx watch scripts/dev.ts",
     "compile": "tsup",
     "eslint": "eslint --ext .js,.cjs,.ts,.cts,.mts,.tsx --no-error-on-unmatched-pattern . --fix",
     "test:playwright": "playwright test --update-snapshots",
@@ -29,6 +29,7 @@
     "@ibm/plex": "^6.3.0",
     "@playwright/test": "^1.37.1",
     "@tailwindcss/forms": "^0.5.3",
+    "@types/cross-spawn": "^6.0.6",
     "@types/express": "^4.17.17",
     "@types/react": "^18.2.21",
     "@types/react-dom": "^18.2.7",
